@@ -18,14 +18,13 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  # sdk: '>=1.9.0 <2.0.0'
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  provider: ^4.3.2+3
+  cupertino_icons: ^1.0.2
+  provider: ^5.0.0-nullsafety.3
   # flutter_localizations package come with flutter sdk, not from pub.dev
   flutter_localizations:
     sdk: flutter
@@ -35,8 +34,8 @@ dependencies:
   video_player: '^1.0.1' # video player controller
   video_thumbnail: ^0.2.5+1 # generates thumbnail from video file or URL
   image_picker: ^0.6.7+21 # enable picking images from the image library, and taking new pictures with the camera.
-  flutter_slidable: ^0.5.7 # enable list item to slide
-  random_string: ^2.1.0 # generating random ascii strings.
+  flutter_slidable: ^0.6.0-nullsafety.0 # enable list item to slide
+  random_string: ^2.2.0-nullsafety # generating random ascii strings.
   # flushbar: # can use this to generate push notification alike UI
 
 dev_dependencies:
