@@ -11,7 +11,7 @@ import 'package:app/data/preferences.dart';
 class AppProviders extends StatelessWidget {
   final Widget child;
 
-  AppProviders({@required this.child});
+  AppProviders({required this.child});
 
   @override
   Widget build(BuildContext context) {
@@ -19,16 +19,14 @@ class AppProviders extends StatelessWidget {
       providers: [
         ChangeNotifierProvider(create: (context) => FileService()),
         ChangeNotifierProvider(
-            create: (context) => AppTheme(AppPreferences().isDarkTheme
-                ? AppTheme.dark()
-                : AppTheme.light())),
+            create: (context) =>
+                AppTheme(AppPreferences().isDarkTheme ? AppTheme.dark() : AppTheme.light())),
         ChangeNotifierProvider(create: (context) => NavigationService()),
         ChangeNotifierProvider(create: (context) => AppPreferences()),
         ChangeNotifierProvider(
           create: (BuildContext context) {
             // return VideoPlayerService.asset();
-            File video =
-                "${FileService.getRootPathSync}/Basic Theming.mp4".getFile;
+            File video = "${FileService.getRootPathSync}/Basic Theming.mp4".getFile;
             return VideoPlayerService.file(video);
           },
         )
