@@ -4,7 +4,7 @@ import 'package:app/services/navigation_service.dart';
 import 'package:app/screens/videos/videos_stack.dart';
 import 'package:app/screens/folders/folders_stack.dart';
 import 'package:app/screens/settings/settings_stack.dart';
-import 'package:app/widgets/tab_navigation_bar.dart';
+import 'package:app/widgets/app_bottom_navigation_bar.dart';
 
 class Root extends StatelessWidget {
   final List<Widget> _stacks = [
@@ -18,14 +18,13 @@ class Root extends StatelessWidget {
     NavigationService _nav = Provider.of<NavigationService>(context);
 
     return WillPopScope(
-      onWillPop: () async =>
-          !await _nav.currentNavigatorkey.currentState.maybePop(),
+      onWillPop: () async => !await _nav.currentNavigatorkey.currentState!.maybePop(),
       child: Scaffold(
         body: IndexedStack(
           index: _nav.currentTabIndex,
           children: this._stacks,
         ),
-        bottomNavigationBar: TabNavigationBar(),
+        bottomNavigationBar: AppBottomNavigationBar(),
       ),
     );
   }
